Clean up some logic
