+++ Sources/MSLNetworking/SessionManager/ApiRouterRequestable.swift
@@ -1,0 +1,16 @@
+import Alamofire
+import Foundation
+
+/// Use this in your dependency injection to obfuscate the underlying session manager
+public protocol ApiRouterRequestable {
+    func request<Response: Decodable>(
+        from request: ApiRouter,
+        using decoder: JSONDecoder
+    ) async throws -> Response
+
+    func request<Response: Decodable, Property: Any>(
+        _ keyPath: KeyPath<Response, Property>,
+        from request: ApiRouter,
+        using decoder: JSONDecoder
+    ) async throws -> Property
+}
